Fix the process for GNU tools on MINIX
This is a fix over commit 0e53925af085b78f9d
On a MINIX station, the tools are not usually built and
on a first-time building of the tree, the fetching script
of texinfo was not triggered in some cases. Let force it.
Reported on minix3 googlegroup by Chris Card.

Change-Id: I8beafdeaec66bb1f1f3250bd64c1e14c0023e9d0